+++ src/Frontend/Modules/Profiles/Actions/ForgotPassword.php
@@ -139,7 +139,7 @@ class ForgotPassword extends FrontendBaseBlock
                 $this->get('mailer')->send($message);
 
                 // redirect
-                $this->redirect(SELF . '?sent=true');
+                $this->redirect(SITE_URL . '/' . $this->URL->getQueryString() . '?sent=true');
             } else {
                 $this->tpl->assign('forgotPasswordHasError', true);
             }