+++ src/Frontend/Modules/Profiles/Actions/Register.php
@@ -186,7 +186,7 @@ class Register extends FrontendBaseBlock
                     $this->get('mailer')->send($message);
 
                     // redirect
-                    $this->redirect(SELF . '?sent=true');
+                    $this->redirect(SITE_URL . '/' . $this->URL->getQueryString() . '?sent=true');
                 } catch (\Exception $e) {
                     // when debugging we need to see the exceptions
                     if (SPOON_DEBUG) {